mediacapture-insertable-streams/MediaStreamTrackProcessor-video.https.html

<!doctype html>
<html>
<head>
  <title>MediaStreamTrackProcessor</title>
  <link rel="help" href="https://w3c.github.io/mediacapture-insertable-streams">
</head>
<body>
<p class="instructions">When prompted, use the accept button to give permission to use your audio and video devices.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that processing captured audio MediaStreamTracks works as expected.</p>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  const track = stream.getTracks()[0];
  const processor = new MediaStreamTrackProcessor({track: track});
  const reader = processor.readable.getReader();
  const doneValue = await reader.read();
  assert_false(doneValue.done)
  assert_true(doneValue.value instanceof VideoFrame);
  doneValue.value.close();
  track.stop();
  assert_false(doneValue.done)
  assert_true(doneValue.value instanceof VideoFrame);
  return reader.closed;
}, "Tests that the reader of a video MediaStreamTrackProcessor produces video frames and is closed on track stop");

promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  const track = stream.getTracks()[0];
  const processor = new MediaStreamTrackProcessor({track: track});
  const worker = new Worker('MediaStreamTrackProcessor-worker.js');
  const promise = new Promise(resolve => {
    worker.onmessage = msg => {
      if (msg.data instanceof VideoFrame) {
        msg.data.close();
        track.stop();
      } else if (msg.data == 'closed') {
        resolve();
      }
    }
  });
  worker.postMessage({readable: processor.readable},
                     [processor.readable]);
  return promise;
}, "Tests that the reader of a video MediaStreamTrackProcessor produces VideoFrame objects and is closed on track stop while running on a worker");
</script>
</body>
</html>
